+++ python/__init__.py
@@ -145,3 +145,5 @@ from .virtual_channel_demultiplexer import virtual_channel_demultiplexer
 from .telemetry_ocf_adder import telemetry_ocf_adder
 from .space_packet_time_stamp_adder import space_packet_time_stamp_adder
 from .telemetry_packet_reconstruction import telemetry_packet_reconstruction
+
+from .smogp_signalling import smogp_signalling